Instrument Operation latencies
The change adds instrumentation to the operation runtimes and has
defined configuration which can be used to send realtime metrics to
wavefront.

Micrometer is the industry standard to instrument code and has a very
good integration with spring boot. The change adds all required
gradle dependencies and the configuration placeholders to enable it to
push to wavefront.

Resolves #1